tikv_alloc/*, server/status_server: move profiler_guard to tikv server
Move profiler_guard from tikv_alloc/profiler_guard to
server/status_server to remove Futures from tikv_alloc dependencies.
