+++ Cargo.toml
@@ -91,6 +91,7 @@ sys-info = "0.5.1"
 indexmap = { version = "1.0", features = ["serde-1"] }
 futures = "0.1"
 futures-cpupool = "0.1"
+futures-locks = "0.1"
 tikv_alloc = { path = "components/tikv_alloc", default-features = false }
 tokio-core = "0.1"
 tokio-timer = "0.2"
