+++ components/tikv_alloc/Cargo.toml
@@ -16,9 +16,6 @@ mem-profiling = ["jemallocator/profiling"]
 [dependencies]
 libc = "0.2"
 log = "0.3"
-lazy_static = "0.2.1"
-futures-locks = "0.1"
-futures = "0.1"
 
 [dev-dependencies]
 tempdir = "0.3"
