+++ components/tikv_alloc/src/lib.rs
@@ -82,15 +82,10 @@
 #[macro_use]
 extern crate log;
 
-#[macro_use]
-extern crate lazy_static;
-
 pub mod error;
 
 #[cfg(not(all(unix, not(fuzzing), feature = "jemalloc")))]
 mod default;
-
-mod profiler_guard;
 
 pub type AllocStats = Vec<(&'static str, usize)>;
 
@@ -106,7 +101,6 @@ mod imp;
 mod imp;
 
 pub use crate::imp::*;
-pub use crate::profiler_guard::*;
 
 #[global_allocator]
 static ALLOC: imp::Allocator = imp::allocator();
