+++ src/server/status_server.rs
@@ -12,11 +12,52 @@ use std::net::SocketAddr;
 use std::str::FromStr;
 
 use super::Result;
-use tikv_alloc;
 use tikv_alloc::error::ProfError;
 use tikv_util::collections::HashMap;
 use tikv_util::metrics::dump;
 use tikv_util::timer::GLOBAL_TIMER_HANDLE;
+
+mod profiler_guard {
+    use tikv_alloc::error::ProfResult;
+    use tikv_alloc::{activate_prof, deactivate_prof};
+
+    use futures::{Future, Poll};
+    use futures_locks::{Mutex, MutexFut, MutexGuard};
+
+    lazy_static! {
+        static ref PROFILER_MUTEX: Mutex<u32> = Mutex::new(0);
+    }
+
+    pub struct ProfGuard(MutexGuard<u32>);
+
+    pub struct ProfLock(MutexFut<u32>);
+
+    impl ProfLock {
+        pub fn new() -> ProfResult<ProfLock> {
+            let guard = PROFILER_MUTEX.lock();
+            match activate_prof() {
+                Ok(_) => Ok(ProfLock(guard)),
+                Err(e) => Err(e),
+            }
+        }
+    }
+
+    impl Drop for ProfGuard {
+        fn drop(&mut self) {
+            match deactivate_prof() {
+                _ => {} // TODO: handle error here
+            }
+        }
+    }
+
+    impl Future for ProfLock {
+        type Item = ProfGuard;
+        type Error = ();
+        fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+            self.0.poll().map(|item| item.map(|guard| ProfGuard(guard)))
+        }
+    }
+}
 
 pub struct StatusServer {
     thread_pool: ThreadPool,
@@ -47,7 +88,7 @@ impl StatusServer {
     }
 
     pub fn dump_prof(seconds: u64) -> Box<dyn Future<Item = Vec<u8>, Error = ProfError> + Send> {
-        let lock = match tikv_alloc::ProfLock::new() {
+        let lock = match profiler_guard::ProfLock::new() {
             Err(e) => return Box::new(err(e)),
             Ok(lock) => lock,
         };
